+++ old_code/get_peers2.py
@@ -1,0 +1,11 @@
+## i got 8 nodes from the call to bootstrap
+## now im gonna make those my initial routing tables
+## im going to call the same function to all 8 and see what i get!!
+## brilliant - im on my way 
+
+K = 8
+#compact node info is K 26byte strings, first 20 is node id, last 6 is ip and port
+#this is the direct response from the get peers calls
+first_nodes = b'\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_\xc6\xe9^\x04GB\xb65B\xc6$\xd2\xaf\xdd\xa2\xed\x0b\r\xbc$\xcf\xccA\x9b\x08_'
+nodes_array = bytearray(first_nodes)
+nodes_list = [[bytes(nodes_array[(26*i):(26*i)+20]), bytes(nodes_array[(26*i)+20:(26*(i+1))])] for i in range(K)]
